Added extra printf statements for console debugging.
Note that the 'deleteNthNodeFromStart' returns the next node in the
list, so while the whole program works, not quite what we are looking
for. Really this should be void - I think I was trying to return the new
list, but this is not what is going on. For that I will need to store
the original head